// frontend/src/app/model/Reassureur.ts
export class Reassureur {
   // reass_id: String;
    reass_code: number;
    reass_codeidentificateur: String;
    reass_codepays: number;
    reass_type: number;
    reass_denomination: String;
    reass_denominationcourt: String;
    reass_adresse1: String;
    reass_adresse2: String;
	reass_telephone1: String;
	reass_telephone2: String;
	reass_email: String;
	reass_datetraite1: String;
	reass_nbretraite: number;
	reass_ca: number;
	reass_commissionrecu: number;
	reass_horsgroupe: String;
	reass_codenationalite: String;
	reass_codeutilisateur: String;
	reass_datemodification: Date;
}
	  